Show commission totals based on current filters
This only shows the net/gross of the current filtered commissions listed for approval.
I had to create a modfied copy of the internal Laravel pagination logic in order to inject the two custom columns, as I wanted to "highjack" the original counting query that is running for the pagination. As it is rather slow, this saves us quite some time.